Remove hard-coded allowed consent codes from ConsentGroup
Because ConsentGroups will be added in the admin interface and not
by users of the app, we don't need to limit the set of allowed codes.
Instead, the models can be created with whatever code you want, but
only a limited set of people (site admins) will be able to create
them. Because we have removed the hard-coded consent codes, we need
to add a field for the short consent status (eg General Research Use"
for GRU). Modify tests, factories, tables, and templates as
appropriate.