Include all urls from the anvil_consortium_manager app
For urls that we want to override with new views (eg WorkspaceDetail
view with WorkspaceDataDetail view), add that url to the main
urls file before the anvil_consortium_manager urls include statement.
That way, Django will find the overriden url first and use that one.

Note that right now there are some views that we *don't* want to
have in the app (e.g., creating a new workspace) and we will need
to direct those to a 404 page later.