Use UploadWorkspace as the workspace data model
Define a custom adapter to specify the model, an accompanying form,
and the custom table to display. Add the form and table as well as
tests for them. Also add tests to make sure that the UploadWorkspace
model is created in the WorkspaceImport and WorkspaceCreate views
from anvil_consortium_manager.