Update for anvil_consortium_manager multiple adapter changes
Update the requirements file to install the newest version of
anvil_consortium_manager. Add required fields to the custom
UploadWorkspace adapter. Add the context processor to settings and
update the required variable setting the adapter. Modify tests as
necessary. Update the base template to include the loop over workspace
types in the navbar.