Change workspace report to have a table by workspace_type
Instead of showing the number of upload workspaces and then the
total number of workspaces for each type, combine those two bullets
into a single table. The new table shows the total number of each
workspace type as well as the number that has been shared with the
consortium. To do this, add a new table and redo the workspace
aggregation queryset to count both the total and the number shared.
Remove the reports.py source file, which is no longer necessary with
the new table.

Note that the new table only shows workspace types where at least
one workspace exists, but I think that makes more sense than showing
0 in both columns (for total number and number shared). It still
includes a rows for workspace_type where a workspace exists, even
if none are shared.