Add a default workspace table for workspace data adapters
This table shows the name of the workspace, the billing project,
the number of groups that it is shared with, and whether it is
shared with the GREGoR consortium via GREGOR_ALL. Also add a table
subclass that adds the "shared with GREGoR?" column. Use this new
default table for all workspace data adapters that don't define
their own class already. Add the "shared with GREGoR?" column to
workspace data tables that define their own table calss, except for
ReleaseWorkspaces (which should never be shared with the
consortium).