Pull in Drupal DSA completed role
For now, we expect that this role will be used to assign the ACM
account link permission. It's possible that this will change in
the future. Note that if we want *everyone* with a Drupal account
to be able to link their AnVIL accounts, we could use the
drupal_machine_name="authorized" scope and map it to a
django_group_name.