Add migrations to populate UploadCycle and set foreign keys
Add three migrations. The first migration adds a nullable upload_cycle
FK to appropriate models. The second migration populates the
UploadCycle model using the set of unique versions that exist in
the UploadWorkspace objects, and sets the upload_cycle foreign
to the appropriate UploadCycle object for UploadWorkspaces,
CombinedConsortiumDataReleaseWorkspaces, and ReleaseWorkspaces.