Modify migration to work on mysql
Change the name of the unique constraints for UploadWorkspace and
ReleaseWorkspace, and move the AddConstraint migration statements
to before the RemoveConstraint statements. This fixes the error in
CI:
```
django.db.utils.OperationalError: (1553, "Cannot drop index 'unique_release_workspace': needed in a foreign key constraint")
```
See also Django bug report (https://code.djangoproject.com/ticket/31335),
which appears to be fixed in later versions but not the version we
are using.