Remove upload_workspace ManyToMany field from combined workspace
This currently is redundant because of the upload_cycle model -- all
upload workspaces from that upload cycle and any previous upload
cycles are included in the combined workspace. We can infer this
implicitly from upload_cycle. If we have a situation where some
workspaces are not included in a combined workspace for some reason,
we can handle this later (either by adding the foreign key back and
writing a migration to handle it, or by adding flags to the upload
workspace).