Update render_is_shared for new sharing structure
Instead of sharing directly with GREGOR_ALL, we will share with
the auth domain and add GREGOR_ALL to the auth domain. Switch
render_is_shared to use the ACM Workspace methods is_shared and
is_in_authorization_domain to determine whether a workspace has
been shared with the consortium.