Allow both direct and indirect dependabot updates
By default all dependencies that are explicitly defined in a
manifest are kept up to date by Dependabot version updates. We got
a security update that reported:
"This is an indirect dependency not explicitly present in any manifest
file, so Dependabot cannot update it."
so try adding dependency-type to the dependabot config.