Add model method to get partner workspaces for an upload cycle
Because the partner workspaces are not linked directly to an upload
cycle, we need to use some logic to get the partner workspaces that
should be included in an upload cycle. For a given research center
and consent group, only the PartnerUploadWorkspace with the highest
version and with a date completed less than the upload cycle end
date will be included for that upload cycle. Use this method in the
UploadCycleDetail view instead of filtering partner workspaces
directly.