Add step that optionally runs pip-compile if specified
If the python version is different than the one running on the
server, we'll need to recompile the pip-tools requirements files.
We don't want to do this for all python versions because it would
not catch situations where we've added a new packages to the
requirements.in files and forgotten to compile them. However, we
need to recompile if we're testing on a new python version because
packages can have different versions of dependencies for different
python versions.