Ignore deprecation warning from third party package
django-environ is throwing a deprecation warning; we can't do anything
about this, so ignore it. Unfortunately we have to pin to a older
version of django-environ because if an issue with reading secret
keys with a hash in them, so hopefully they will fix it before we
need to switch to python 3.14.