Change "release" to "release prep" in user-facing text
These are not the actual workspaces that are getting released; instead
they are workspaces where we are preparing the release. (This changed
since they were initially coded.) To help prevent confusion from users,
show "release prep" instead of "release" in text that users see. Do
not make any changes to the underlying structure of the code or
models, because we are not tracking the actual release workspaces
in ACM.