Add second exception to try/except block
The AnVILGroupNotFound exception is returned when you try to share
a workspace with a group, but the group does not exist in AnVIL.
The view that resolves the audit should handle this like an API
error (even though the return code is success - 200 - it is not
actually a success). Without this exception in the try/except block,
the htmx functionality did not return either "Handled!" or "Error!",
which is confusing to the user.