Do not schedule combine-prs workflows
It is better to review the dependabot PRs, close any that we do
not want included in the combined update, and then run the combine-prs
workflow manually (via workflow dispatch). This way we won't
inadvertently get updates that we don't want (e.g., a passing PR that
updates to django-5.0).