Update CI to fail if no files are found
The coverage job is failing because github added a breaking change
in v4.4 of the upload-artifact job such that hidden files were no
longer uploaded by default. This breaks the current coverage setup
where the coverage file is hidden (.coverage). This change causes
the upload-artifact job to fail if no files are found, which is
really what we want. With this change, I expect all pytest jobs
to fail at the step of uploading the coverage file.