Add basic ability to update contributing workspaces
Add a form and a view for updating contributing workspaces for a release
prep workspace. The page is not pretty yet, and the form won't work
well when there are a large number of workespaces (like in prod) but
that can be updated later.

Still needs to be prettified, and widgets need to be updated