Fix form to use WorkspaceAutocompleteByType view
The contributing workspaces form was using the WorkspaceAutocomplete
view, which returned Workspace objects. We need it to operate on
workspace data objects (UploadWorkspaces or DCCProcessedDataWorkspaces).
To accomplish this, I needed to add a new namespace and url pattern
for the WorkspaceAutoCompleteByType view where the workspace_type url
argument is passed. I initially tried calling reverse in the form widget
ModelSelect2Multiple call as suggested by this issue/PR fix:
but it caused a circular import. Adding new url patterns is a
(possibly?) hacky but functional workaround.